# openapi.json
{
  "components": {
    "schemas": {
...
      "ConsistencyRow": {
        "properties": {
          "avg_score": {
            "title": "Avg Score",
            "type": "number"
          },
          "manager_id": {
            "title": "Manager Id",
            "type": "integer"
          },
          "manager_name": {
            "title": "Manager Name",
            "type": "string"
          },
          "max_score": {
            "title": "Max Score",
            "type": "number"
          },
          "min_score": {
            "title": "Min Score",
            "type": "number"
          },
          "std_dev": {
            "title": "Std Dev",
            "type": "number"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "manager_id",
          "manager_name",
          "year",
          "avg_score",
          "std_dev",
          "min_score",
          "max_score"
        ],
        "title": "ConsistencyRow",
        "type": "object"
      },
      "ConsolationRow": {
        "properties": {
          "consolation_avg_points": {
            "title": "Consolation Avg Points",
            "type": "number"
          },
          "consolation_games": {
            "title": "Consolation Games",
            "type": "integer"
          },
          "consolation_losses": {
            "title": "Consolation Losses",
            "type": "integer"
          },
          "consolation_win_pct": {
            "title": "Consolation Win Pct",
            "type": "number"
          },
          "consolation_wins": {
            "title": "Consolation Wins",
            "type": "integer"
          },
          "manager_id": {
            "title": "Manager Id",
            "type": "integer"
          },
          "manager_name": {
            "title": "Manager Name",
            "type": "string"
          },
          "times_missed_playoffs": {
            "title": "Times Missed Playoffs",
            "type": "integer"
          }
        },
        "required": [
          "manager_id",
          "manager_name",
          "times_missed_playoffs",
          "consolation_games",
          "consolation_wins",
          "consolation_losses",
          "consolation_win_pct",
          "consolation_avg_points"
        ],
        "title": "ConsolationRow",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LuckIndexRow": {
        "properties": {
          "actual_wins": {
            "title": "Actual Wins",
            "type": "integer"
          },
          "expected_wins": {
            "title": "Expected Wins",
            "type": "number"
          },
          "luck_score": {
            "title": "Luck Score",
            "type": "number"
          },
          "manager_id": {
            "title": "Manager Id",
            "type": "integer"
          },
          "manager_name": {
            "title": "Manager Name",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "manager_id",
          "manager_name",
          "year",
          "actual_wins",
          "expected_wins",
          "luck_score"
        ],
        "title": "LuckIndexRow",
        "type": "object"
      },
...
      "ManagerTierRow": {
        "properties": {
          "avg_ppg": {
            "title": "Avg Ppg",
            "type": "number"
          },
          "championships": {
            "title": "Championships",
            "type": "integer"
          },
          "composite_score": {
            "title": "Composite Score",
            "type": "number"
          },
          "consistency_score": {
            "title": "Consistency Score",
            "type": "number"
          },
          "dimension_scores": {
            "$ref": "#/components/schemas/ManagerTierDimensions"
          },
          "expected_win_pct": {
            "title": "Expected Win Pct",
            "type": "number"
          },
          "manager_id": {
            "title": "Manager Id",
            "type": "integer"
          },
          "manager_name": {
            "title": "Manager Name",
            "type": "string"
          },
          "playoff_rate": {
            "title": "Playoff Rate",
            "type": "number"
          },
          "seasons_played": {
            "title": "Seasons Played",
            "type": "integer"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          },
          "win_pct": {
            "title": "Win Pct",
            "type": "number"
          }
        },
        "required": [
          "manager_id",
          "manager_name",
          "tier",
          "composite_score",
          "win_pct",
          "avg_ppg",
          "expected_win_pct",
          "championships",
          "playoff_rate",
          "consistency_score",
          "seasons_played",
          "dimension_scores"
        ],
        "title": "ManagerTierRow",
        "type": "object"
      },
...
      "PlayoffPerformanceRow": {
        "properties": {
          "delta_avg_pts": {
            "title": "Delta Avg Pts",
            "type": "number"
          },
          "delta_win_pct": {
            "title": "Delta Win Pct",
            "type": "number"
          },
          "manager_id": {
            "title": "Manager Id",
            "type": "integer"
          },
          "manager_name": {
            "title": "Manager Name",
            "type": "string"
          },
          "playoff_avg_pts": {
            "title": "Playoff Avg Pts",
            "type": "number"
          },
          "playoff_games": {
            "title": "Playoff Games",
            "type": "integer"
          },
          "playoff_losses": {
            "title": "Playoff Losses",
            "type": "integer"
          },
          "playoff_win_pct": {
            "title": "Playoff Win Pct",
            "type": "number"
          },
          "playoff_wins": {
            "title": "Playoff Wins",
            "type": "integer"
          },
          "reg_avg_pts": {
            "title": "Reg Avg Pts",
            "type": "number"
          },
          "reg_games": {
            "title": "Reg Games",
            "type": "integer"
          },
          "reg_losses": {
            "title": "Reg Losses",
            "type": "integer"
          },
          "reg_win_pct": {
            "title": "Reg Win Pct",
            "type": "number"
          },
          "reg_wins": {
            "title": "Reg Wins",
            "type": "integer"
          }
        },
        "required": [
          "manager_id",
          "manager_name",
          "reg_wins",
          "reg_losses",
          "reg_win_pct",
          "reg_avg_pts",
          "reg_games",
          "playoff_wins",
          "playoff_losses",
          "playoff_win_pct",
          "playoff_avg_pts",
          "playoff_games",
          "delta_win_pct",
          "delta_avg_pts"
        ],
        "title": "PlayoffPerformanceRow",
        "type": "object"
      },
...
      "PowerRankingRow": {
        "properties": {
          "dimensions": {
            "$ref": "#/components/schemas/PowerDimensions"
          },
          "manager_id": {
            "title": "Manager Id",
            "type": "integer"
          },
          "manager_name": {
            "title": "Manager Name",
            "type": "string"
          },
          "overall_score": {
            "title": "Overall Score",
            "type": "number"
          }
        },
        "required": [
          "manager_id",
          "manager_name",
          "overall_score",
          "dimensions"
        ],
        "title": "PowerRankingRow",
        "type": "object"
      },
      "ProjectionRow": {
        "properties": {
          "avg_actual": {
            "title": "Avg Actual",
            "type": "number"
          },
          "avg_diff": {
            "title": "Avg Diff",
            "type": "number"
          },
          "avg_projected": {
            "title": "Avg Projected",
            "type": "number"
          },
          "beat_projection_pct": {
            "title": "Beat Projection Pct",
            "type": "number"
          },
          "manager_id": {
            "title": "Manager Id",
            "type": "integer"
          },
          "manager_name": {
            "title": "Manager Name",
            "type": "string"
          },
          "std_dev_diff": {
            "title": "Std Dev Diff",
            "type": "number"
          },
          "total_over": {
            "title": "Total Over",
            "type": "number"
          },
          "weeks": {
            "title": "Weeks",
            "type": "integer"
          }
        },
        "required": [
          "manager_id",
          "manager_name",
          "weeks",
          "avg_actual",
          "avg_projected",
          "avg_diff",
          "beat_projection_pct",
          "std_dev_diff",
          "total_over"
        ],
        "title": "ProjectionRow",
        "type": "object"
      },
      "RivalryDetail": {
        "properties": {
          "biggest_blowout": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RivalryMatchup"
              },
              {
                "type": "null"
              }
            ]
          },
          "closest_game": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RivalryMatchup"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_streak": {
            "$ref": "#/components/schemas/RivalryStreak"
          },
          "longest_streak": {
            "$ref": "#/components/schemas/RivalryStreak"
          },
          "manager_a_id": {
            "title": "Manager A Id",
            "type": "integer"
          },
          "manager_a_name": {
            "title": "Manager A Name",
            "type": "string"
          },
          "manager_b_id": {
            "title": "Manager B Id",
            "type": "integer"
          },
          "manager_b_name": {
            "title": "Manager B Name",
            "type": "string"
          },
          "matchups": {
            "items": {
              "$ref": "#/components/schemas/RivalryMatchup"
            },
            "title": "Matchups",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/RivalrySummary"
          },
          "trends": {
            "items": {
              "$ref": "#/components/schemas/RivalryTrend"
            },
            "title": "Trends",
            "type": "array"
          }
        },
        "required": [
          "manager_a_id",
          "manager_a_name",
          "manager_b_id",
          "manager_b_name",
          "summary",
          "matchups",
          "trends",
          "current_streak",
          "longest_streak",
          "biggest_blowout",
          "closest_game"
        ],
        "title": "RivalryDetail",
        "type": "object"
      },
...
      "SeasonAwards": {
        "properties": {
          "awards": {
            "items": {
              "$ref": "#/components/schemas/Award"
            },
            "title": "Awards",
            "type": "array"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "year",
          "awards"
        ],
        "title": "SeasonAwards",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Fantasy Football Dashboard API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/stats/awards": {
      "get": {
        "operationId": "awards_api_stats_awards_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonAwards"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Awards",
        "tags": [
          "stats"
        ]
      }
    },
    "/api/stats/consistency": {
      "get": {
        "operationId": "consistency_api_stats_consistency_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsistencyRow"
                  },
                  "title": "Response Consistency Api Stats Consistency Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consistency",
        "tags": [
          "stats"
        ]
      }
    },
    "/api/stats/consolation": {
      "get": {
        "operationId": "consolation_bracket_api_stats_consolation_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsolationRow"
                  },
                  "title": "Response Consolation Bracket Api Stats Consolation Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consolation Bracket",
        "tags": [
          "stats"
        ]
      }
    },
...
    "/api/stats/luck-index": {
      "get": {
        "operationId": "luck_index_api_stats_luck_index_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LuckIndexRow"
                  },
                  "title": "Response Luck Index Api Stats Luck Index Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Luck Index",
        "tags": [
          "stats"
        ]
      }
    },
    "/api/stats/manager-tiers": {
      "get": {
        "operationId": "manager_tiers_api_stats_manager_tiers_get",
        "parameters": [
          {
            "in": "query",
            "name": "year_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year Start"
            }
          },
          {
            "in": "query",
            "name": "year_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ManagerTierRow"
                  },
                  "title": "Response Manager Tiers Api Stats Manager Tiers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Manager Tiers",
        "tags": [
          "stats"
        ]
      }
    },
    "/api/stats/playoff-performance": {
      "get": {
        "operationId": "playoff_performance_api_stats_playoff_performance_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlayoffPerformanceRow"
                  },
                  "title": "Response Playoff Performance Api Stats Playoff Performance Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Playoff Performance",
        "tags": [
          "stats"
        ]
      }
    },
...
    "/api/stats/power-rankings": {
      "get": {
        "operationId": "power_rankings_api_stats_power_rankings_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PowerRankingRow"
                  },
                  "title": "Response Power Rankings Api Stats Power Rankings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Power Rankings",
        "tags": [
          "stats"
        ]
      }
    },
    "/api/stats/projection-performance": {
      "get": {
        "operationId": "projection_performance_api_stats_projection_performance_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectionRow"
                  },
                  "title": "Response Projection Performance Api Stats Projection Performance Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Projection Performance",
        "tags": [
          "stats"
        ]
      }
    },
    "/api/stats/rivalry": {
      "get": {
        "operationId": "rivalry_api_stats_rivalry_get",
        "parameters": [
          {
            "in": "query",
            "name": "manager_a",
            "required": true,
            "schema": {
              "title": "Manager A",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "manager_b",
            "required": true,
            "schema": {
              "title": "Manager B",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RivalryDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rivalry",
        "tags": [
          "stats"
        ]
      }
    },
...
    "/api/stats/strength-of-schedule": {
      "get": {
        "operationId": "strength_of_schedule_api_stats_strength_of_schedule_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrengthOfScheduleRow"
                  },
                  "title": "Response Strength Of Schedule Api Stats Strength Of Schedule Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Strength Of Schedule",
        "tags": [
          "stats"
        ]
      }
    },
...
    "/api/stats/trophy-case/{manager_id}": {
      "get": {
        "operationId": "trophy_case_api_stats_trophy_case__manager_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "manager_id",
            "required": true,
            "schema": {
              "title": "Manager Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrophyCase"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trophy Case",
        "tags": [
          "stats"
        ]
      }
    },
...
    "/api/stats/weekly-records": {
      "get": {
        "operationId": "weekly_records_api_stats_weekly_records_get",
        "parameters": [
          {
            "in": "query",
            "name": "top_n",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Top N",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyRecords"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Weekly Records",
        "tags": [
          "stats"
        ]
      }
    },
    "/api/stats/win-margins": {
      "get": {
        "operationId": "win_margins_api_stats_win_margins_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WinMarginRow"
                  },
                  "title": "Response Win Margins Api Stats Win Margins Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Win Margins",
        "tags": [
          "stats"
        ]
      }
    },
...
  }
}